Restore fetch version number error scenario behavior
