+++ spec/appcenter_fetch_app_action_spec.rb
@@ -1,3 +1,5 @@
+require_relative 'stubs'
+
 def stub_get_app_not_found(status)
   not_found_json = JSON.parse(File.read("spec/fixtures/releases/not_found.json"))
   stub_request(:get, "https://api.appcenter.ms/v0.1/apps/owner-name/App-Name")
