+++ spec/fixtures/releases/no_releases.json
@@ -1,0 +1,4 @@
+{
+  "code": "not_found",
+  "message": "Error: No release found for appId: aaaaaaaa-1111-1aa1-1aa1-1111aaaa1111, userId: aaaaaaaa-1111-1aa1-1aa1-1111aaaa1111"
+}
